// src/app/models/app-routes.ts
export enum AppRoutes {
    ROOT = '/',
    REGISTER = 'register',
    INVITE = 'invite',
    LOGIN = 'login',
    FORGOT_PASS = 'forgotpassword',
    OTP_VERIFY = 'otpverification',
    OTP_VERIFY_INVITED_USER = 'otpverification',
    RESET_PASS = 'resetpassword',
    APPOINTMENT = 'appointment',
    HISTORY = 'history',
    USER_REG_F = 'userregistrationform',
    FEEDBACK_FORM='feedbackform',
    USER_REG_S = 'userregistrationformsecond',

    BOOKING_TIMINGS='bookingtimings',

    PROFILE = 'profile',
    REGISTER_CENTER = 'registercenter',
    PRIVACY_POLICY = 'privacypolicy',
    TERMS_CON = 'termsandconditions',
    NOTIFICATION='notifications'
}